tflite ex works somewhat, cleaned up
